+++ graylog2-server/src/main/java/org/graylog/security/rest/EntitySharesResource.java
@@ -16,10 +16,10 @@
  */
 package org.graylog.security.rest;
 
+import com.google.inject.Inject;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import jakarta.inject.Inject;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
