+++ graylog2-server/src/test/java/org/graylog/events/processor/aggregation/AggregationEventProcessorConfigTest.java
@@ -41,7 +41,7 @@ import org.graylog.plugins.views.search.searchtypes.pivot.series.Max;
 import org.graylog.plugins.views.search.searchtypes.pivot.series.Sum;
 import org.graylog.scheduler.schedule.IntervalJobSchedule;
 import org.graylog.security.entities.EntityOwnershipService;
-import org.graylog.security.shares.EntitySharesService;
+import org.graylog.security.shares.EntitySharesServiceImpl;
 import org.graylog.testing.mongodb.MongoDBFixtures;
 import org.graylog.testing.mongodb.MongoDBInstance;
 import org.graylog2.bindings.providers.MongoJackObjectMapperProvider;
@@ -99,7 +99,7 @@ public class AggregationEventProcessorConfigTest {
         final MongoJackObjectMapperProvider mapperProvider = new MongoJackObjectMapperProvider(objectMapper);
         final MongoCollections mongoCollections = new MongoCollections(mapperProvider, mongodb.mongoConnection());
         this.dbService = new DBEventDefinitionService(mongoCollections, stateService,
-                mock(EntityOwnershipService.class), null, new IgnoreSearchFilters(), mock(EntitySharesService.class));
+                mock(EntityOwnershipService.class), null, new IgnoreSearchFilters(), mock(EntitySharesServiceImpl.class));
         this.clock = new JobSchedulerTestClock(DateTime.now(DateTimeZone.UTC));
     }
