+++ graylog2-server/src/test/java/org/graylog/plugins/views/search/views/ViewServiceTest.java
@@ -22,7 +22,7 @@ import com.google.common.collect.ImmutableSet;
 import org.graylog.plugins.views.search.permissions.SearchUser;
 import org.graylog.plugins.views.search.rest.TestSearchUser;
 import org.graylog.security.entities.EntityOwnershipService;
-import org.graylog.security.shares.EntitySharesService;
+import org.graylog.security.shares.EntitySharesServiceImpl;
 import org.graylog.testing.ObjectMapperExtension;
 import org.graylog.testing.mongodb.MongoDBExtension;
 import org.graylog.testing.mongodb.MongoDBTestService;
@@ -77,7 +77,7 @@ public class ViewServiceTest {
                 mock(EntityOwnershipService.class),
                 mock(ViewSummaryService.class),
                 mongoCollections,
-                mock(EntitySharesService.class));
+                mock(EntitySharesServiceImpl.class));
         this.searchUser = TestSearchUser.builder().build();
     }
