+++ graylog2-server/src/test/java/org/graylog2/contentpacks/facades/StreamCatalogTest.java
@@ -23,7 +23,7 @@ import com.google.common.collect.ImmutableSet;
 import org.bson.types.ObjectId;
 import org.graylog.events.legacy.V20190722150700_LegacyAlertConditionMigration;
 import org.graylog.security.entities.EntityOwnershipService;
-import org.graylog.security.shares.EntitySharesService;
+import org.graylog.security.shares.EntitySharesServiceImpl;
 import org.graylog.testing.mongodb.MongoDBFixtures;
 import org.graylog.testing.mongodb.MongoDBInstance;
 import org.graylog2.contentpacks.EntityDescriptorIds;
@@ -91,7 +91,7 @@ public class StreamCatalogTest {
     @Mock
     private EntityOwnershipService entityOwnershipService;
     @Mock
-    private EntitySharesService entitySharesService;
+    private EntitySharesServiceImpl entitySharesService;
     @Mock
     private UserService userService;
