+++ graylog2-server/src/test/java/org/graylog2/streams/StreamServiceImplTest.java
@@ -19,7 +19,7 @@ package org.graylog2.streams;
 import com.google.common.collect.ImmutableSet;
 import org.bson.types.ObjectId;
 import org.graylog.security.entities.EntityOwnershipService;
-import org.graylog.security.shares.EntitySharesService;
+import org.graylog.security.shares.EntitySharesServiceImpl;
 import org.graylog.testing.mongodb.MongoDBFixtures;
 import org.graylog.testing.mongodb.MongoDBInstance;
 import org.graylog2.database.NotFoundException;
@@ -61,7 +61,7 @@ public class StreamServiceImplTest {
     @Mock
     private EntityOwnershipService entityOwnershipService;
     @Mock
-    private EntitySharesService entitySharesService;
+    private EntitySharesServiceImpl entitySharesService;
 
     private StreamService streamService;
